% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Institution.R
\name{addSingleContract}
\alias{addSingleContract}
\title{addSingleContract}
\usage{
addSingleContract(inst)
}
\description{
assignSingleContract(institution, ptf) assigns a given Portfolio <ptf>
to the respective leaf of the institution tree while converting the contracts
of the given Portfolio <ptf> into a data.frame first.
}
