% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annuities}
\alias{annuities}
\title{annuities}
\format{
A data frame with 27 rows and 36 variables:
\describe{
  \item{\code{calendar}}{character Calendar defines the non-working days which affect the dates of contract events.}
  \item{\code{businessDayConvention}}{character Business Day Convention is linked to the calendar which defines working and non-working days. On this basis the BDC defines how to treat them (no shift / shift of calculation).}
  \item{\code{endOfMonthConvention}}{character End of Month Convention defines handling of the end of month in case of anchoor dates.}
  \item{\code{contractType}}{character The Contract Type defines the cash flow generating pattern of a contract.}
  \item{\code{statusDate}}{double Status Date holds the date per which all attributes of the record were updated.}
  \item{\code{contractRole}}{character Contract Role defines if the contract is an asset or liability, a long or short position.}
  \item{\code{legalEntityIDRecordCreator}}{character Identifies the legal entity creating the contract record.}
  \item{\code{contractID}}{character Unique identifier of a contract.}
  \item{\code{legalEntityIDCounterparty}}{character Identifies the counterparty of the contract}
  \item{\code{cycleAnchorDateOfInterestPayment}}{double Date from which the interest payment date schedule is calculated according to the cycle length. The first interest payment event takes place on this anchor.}
  \item{\code{cycleOfInterestPayment}}{character Cycle according to which the interest payment date schedule will be calculated.}
  \item{\code{nominalInterestRate}}{double The nominal interest rate which will be used to calculate accruals and the next interest payment at the next interest payment date.}
  \item{\code{dayCountConvention}}{double Method defining how days are counted between two dates. This finally defines the year fraction in accrual calculations.}
  \item{\code{accruedInterest}}{character Accrued interest as per status date.}
  \item{\code{cyclePointOfInterestPayment}}{character Defines if interest is paid at the beginning or the end of the cycle.}
  \item{\code{currency}}{character The currency of the cash flows.}
  \item{\code{amortizationDate}}{double Amortization Date is used to calculate the annuity amounts.}
  \item{\code{contractDealDate}}{double Contract Deal Date signifies the origination of the contract where an agreement between the customer and the bank has been settled.}
  \item{\code{initialExchangeDate}}{double Date of the initial cash flow for Maturity and Non-Maturity contracts. It also coincides with the beginning of interest accrual calculation.}
  \item{\code{premiumDiscountAtIED}}{double Total original premium or discount. Negative value for discount and positive for premium.}
  \item{\code{maturityDate}}{double Marks the contractual end of the lifecycle of a contract. Generally, date of the last cash flows.}
  \item{\code{notionalPrincipal}}{double Current nominal value of the contract.}
  \item{\code{cycleAnchorDateOfPrincipalRedemption}}{double Date from which the principal payment date schedule is calculated according to the cycle length. The first principal payment event takes place on this anchor.}
  \item{\code{cycleOfPrincipalRedemption}}{character Cycle according to which the interest payment date schedule will be calculated.}
  \item{\code{nextPrincipalRedemptionPayment}}{character Amount of principal that will be paid during the redemption cycle at the next payment date.}
  \item{\code{terminationDate}}{character If a contract is sold before MD (for example a bond on the secondary market) this date has to be set.}
  \item{\code{cycleAnchorDateOfRateReset}}{logical Date from which the rate reset date schedule is calculated according to the cycle length. The first rate reset event takes place on this anchor.}
  \item{\code{cycleOfRateReset}}{logical Cycle according to which the rate reset date schedule will be calculated.}
  \item{\code{rateSpread}}{double Interest rate spread. A typical rate resetting rule is LIBOR plus x basis point where x represents the interest rate spread.}
  \item{\code{marketObjectCodeOfRateReset}}{logical Is pointing to the interest rate driver (MarketObject) used for rate reset uniquely. Unique codes for market objects must be used.}
  \item{\code{cyclePointOfRateReset}}{character Defines if rate reset happens at the beginning or at the end of the cycle.}
  \item{\code{rateMultiplier}}{double Interest rate multiplier.}
  \item{\code{description}}{character Description of the financial contract.}
  \item{\code{contrStrucObj.marketObjectCode}}{character tbd}
  \item{\code{contrStruc.referenceType}}{character tbd}
  \item{\code{contrStruc.referenceRole}}{character tbd} 
}
}
\source{
\url{https://www.actusfrf.org/dictionary/}
}
\usage{
annuities
}
\description{
This datasheet holds predefined financial contracts forming a portfolio of annuities for an institution.
}
\keyword{datasets}
