% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Default.R, R/Institution.R
\docType{methods}
\name{default}
\alias{default}
\alias{generateDefaultContracts}
\alias{generateDefaultContracts,ContractType,list,character,data.frame,numeric-method}
\alias{default,Node,list,character,numeric-method}
\alias{determineDefault}
\title{Default}
\usage{
default(object, defaults, from, recoveryRate, ...)

generateDefaultContracts(object, defaults, from, ctr, recoveryRate, ...)

\S4method{generateDefaultContracts}{ContractType,list,character,data.frame,numeric}(object, defaults, from, ctr, recoveryRate)

\S4method{default}{Node,list,character,numeric}(object, defaults, from, recoveryRate)

determineDefault(node, defaults, from, recoveryRate)
}
\arguments{
\item{object}{object The \code{Institution}-object for which the contracts indicate a default risk}

\item{defaults}{A list representing the default curves for the different economic sectors}

\item{from}{A character defining a date for the default simulation}
}
\description{
Default
}
