% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findNodeByName}
\alias{findNodeByName}
\title{********************************************************
findNodeByName(node, name)
 get the node object of a data.tree with a given string
********************************************************}
\usage{
findNodeByName(node, name)
}
\description{
********************************************************
findNodeByName(node, name)
 get the node object of a data.tree with a given string
********************************************************
}
