% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Institution.R
\name{getContractsAsDataFrames}
\alias{getContractsAsDataFrames}
\title{getContractsAsDataFrames}
\usage{
getContractsAsDataFrames(institution, node, ...)
}
\description{
getContractsAsDataFrames(institution, node) converts each contract 
under the given node to a data.frame
}
