% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{operations}
\alias{operations}
\title{operations}
\format{
A data frame with 10 rows and 14 variables:
\describe{
  \item{\code{node}}{character The node defines the leaf of the financial instrument.}
  \item{\code{contractType}}{character The Contract Type defines the cash flow generating pattern of a contract.}
  \item{\code{contractID}}{character Unique identifier of a contract.}
  \item{\code{contractRole}}{character Contract Role defines if the contract is an asset or liability, a long or short position.}
  \item{\code{currency}}{character The currency of the cash flows.}
  \item{\code{notionalPrincipal}}{character Current nominal value of the contract.}
  \item{\code{nominalInterestRate}}{character The nominal interest rate which will be used to calculate accruals and the next interest payment at the next interest payment date.}
  \item{\code{initialExchangeDate}}{character Date of the initial cash flow for Maturity and Non-Maturity contracts. It also coincides with the beginning of interest accrual calculation.}
  \item{\code{maturityDate}}{character Marks the contractual end of the lifecycle of a contract. Generally, date of the last cash flows.}
  \item{\code{repetition}}{numeric}
  \item{\code{frequency}}{numeric}
  \item{\code{times}}{numeric}
  \item{\code{inverted}}{logical}
  \item{\code{description}}{character Description of the operation contract.}
  
}
}
\source{
\url{https://www.actusfrf.org/dictionary/}
}
\usage{
operations
}
\description{
This datasheet holds predefined financial contracts forming a portfolio of operations for an institution.
}
\keyword{datasets}
