% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yieldCurves}
\alias{yieldCurves}
\title{yieldCurves}
\format{
A data frame with 2 rows and 13 variables:
\describe{
  \item{\code{rfType}}{character Indicates risk factor type: in this stage only YieldCurve supported.}
  \item{\code{label}}{character Adds a label/name to the risk factor object.}
  \item{\code{referenceDate}}{double Adds valid from date to the risk factor object.}
  \item{\code{tenor.1}}{character tenor.n defines the time span or length of time of rate.n to be valid }
  \item{\code{rate.1}}{double rate.n defines the interest rate for given tenor.n.}
  \item{\code{tenor.2}}{character tenor.n defines the time span or length of time of rate.n to be valid}
  \item{\code{rate.2}}{double rate.n defines the interest rate for given tenor.n.}
  \item{\code{tenor.3}}{character tenor.n defines the time span or length of time of rate.n to be valid}
  \item{\code{rate.3}}{double rate.n defines the interest rate for given tenor.n.}
  \item{\code{tenor.4}}{character tenor.n defines the time span or length of time of rate.n to be valid}
  \item{\code{rate.4}}{double rate.n defines the interest rate for given tenor.n.}
  \item{\code{tenor.5}}{character tenor.n defines the time span or length of time of rate.n to be valid}
  \item{\code{rate.5}}{double rate.n defines the interest rate for given tenor.n.} 
}
}
\source{
\url{https://github.com/hamzavig/FRSA/}
}
\usage{
yieldCurves
}
\description{
This datasheet holds predefined Yield Curves.
}
\keyword{datasets}
